pulled out tree component for reuse
